allow keys to be entered in hex-encoded format
If a user needs to handle arbitrary keys, the existing mechanism of
directly using `fgets` is not sufficient.  It will terminate input on
newline, and is also unable to handle 0 (NUL) bytes in input.  To work
around this, this patch adds a `--hex` option to several operations
that instructs the password-reading code to expect a hex-encoded
string.  It will then decode the string into the original bytes before
use.